fixed the ci/cd errors
